+++ src/App.js
@@ -1,12 +1,28 @@
 import React from 'react';
 import './App.css';
-import Home from './components/home/Home'
+
+import { Switch, Route } from 'react-router-dom'; // HERE
+
+import BeersList from './components/BeersList';
+import RandomBeer from './components/RandomBeer';
+import NewBeer from './components/NewBeer';
+
+import Home from './components/Home'
+import Header from './components/header/Header'
 
 function App() {
   return (
     <div className="App">
-      <Home />
-    </div>
+    <Header />
+
+    <Switch> 
+      <Route exact path="/" component={Home}/>
+      <Route exact path="/beers" component={BeersList}/>
+      <Route exact path="/random-beer" component={RandomBeer} />
+      <Route exact path="/new-beer" component={NewBeer} />
+    </Switch>
+
+  </div>
   );
 }
