+++ src/components/BeersList.js
@@ -1,10 +1,11 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 class BeersList extends React.Component {
   render () {
     return (
       <>
-        
+        <h1><Link to="/beers">All Beers</Link></h1>
       </>
     ) 
   }
