+++ src/components/assets/beers.png
